Greatly simplify types
