--- src/types/native/types/NativeType.ts
@@ -1,1 +1,0 @@
-export type NativeType = number | string | boolean | symbol | object | unknown;
