+++ test/types/tests/error/testInferenceOnErrorTypes.ts
@@ -1,0 +1,9 @@
+import { typeIt } from "../../../../src/utility/core/typeIt";
+
+export function testInferenceOnErrorTypes() {
+  const value = new Error();
+  const type = Error;
+  if (typeIt(value, type)) {
+    value.message;
+  }
+}
